admin can add rights

+++ app/models/SubmissionsModel.php
@@ -3,7 +3,6 @@
 
 class SubmissionsModel extends MenuModel
 {
-    private $submissions;
     public function __construct()
     {
         parent::__construct();
@@ -11,7 +10,12 @@ class SubmissionsModel extends MenuModel
     }
 
     public function DeleteSubmission(){
+        if(!ctype_digit(Utils::SafePost('submissionToDeleteId')))
+        {
+            return false;
+        }
         $submissionId = intval(Utils::SafePost('submissionToDeleteId'));
+
         if(Validator::IsUsersSubmission($submissionId)){
             $this->deletePdf($submissionId);
 
@@ -23,10 +27,10 @@ class SubmissionsModel extends MenuModel
     }
 
     private function loadSubmissions(){
-        $this->submissions = Database::GetInstance()->Select(
+        $submissions = Database::GetInstance()->Select(
             SUBMISSIONS_TABLE, '*', 'userId = ?', [Utils::UserId()]);
 
-        $this->data['submissions'] = $this->submissions;
+        $this->data['submissions'] = $submissions;
     }
 
     private function deletePdf($submissionId){
